Update 52.md
Based on comments (thank you!), I rewrote this NIP and focused purely on badges, which was just one of the use cases under "Issued Claims". This greatly simplified the NIP. 

I also specified the data schemas for two different badge types to be part of the protocol itself, versus leaving it open. A bunch of stuff, like the Refresh Claim Request event or requiring event treated as parameterized replaceable event were removed as not necessary for just the badges use case.

There are probably more badge types that would be useful.