NIP-07 friendly description
When people use web clients for the first time, many are redirected to this page to learn about signing in with a browser extension. This page keeps a good list of options, but it is technical.
Adding a "Nostr How To" here may make it easier for non-devs to understand NIP-07.