Adding ability to have expiring mutelist items
I’m working on implementing mutelists in Damus, and users are asking for the ability to have mute list items expire.

After talking with @jb55, one important design decision here is that tags in a kind:10000 mute list do not have the same parameters as kind1 events. For example, we do not include relay URL or petname as those do not make sense in this context. Instead the 3rd element (index: 2) will be the expiration time.

This is something that I’m currently developing for Damus, and wanted to submit this proposal before I get too far into the implementation to gather feedback.

Signed-off-by: Charlie Fish <contact@charlie.fish>