More precisely document since/until bounds. (#1284)
* More precisely document since/until bounds.

Lower in this document, this is expressed as "since <= created_at <= until".

But the English description said that these values had to be "newer than" or "older than".

These two were in conflict. I assume the inclusive behavior is the intended one, since that documentation was added later to clarify. This just makes both cases match.

* Update 01.md

Co-authored-by: Alex Gleason <alex@alexgleason.me>

* Update 01.md

Co-authored-by: Alex Gleason <alex@alexgleason.me>

---------

Co-authored-by: Alex Gleason <alex@alexgleason.me>