Update 11.md with banner field
Added banner relay metadata field. 

If nostr relay management is not to be relegated to PhD or protocol developers only, an effort should be made by relay owners to communicate with non-dev non-technical nostr end users. A banner is a visual representation of the relay. It should aim to visually communicate the brand of the relay.