Add optional arg to nostr.getPublicKey()
If login is optional to an application,
but we want to use the pubkey if the user
had already granted permission to the app,
we could use nostr.getPublicKey({optional: true}).

This will let the user read the app's landing page first,
without being interrupted by a popup window. After that
the user can still press the login button, which calls
nostr.getPublicKey() and prompts the user for permission.