Consistent 'method' naming for NIP 47(NWC) Response
Add consistency to the event structure in (NWC). 

Currently, requests and responses have different keys to identify their method type, which may cause confusion or misunderstandings. 

This change aims to standardize the key names across all request and response events, making it easier for developers to understand and implement the protocol