NIP-34!: use NIP-22 threading for status
deprecate NIP-10 style threading and replace with NIP-22 style for
additional status.

merged patch event are referenced using the `q` tag even though the
they are not included in `content`. this is because NIP-22 only
supports the use of `e` reference for the parent event id.

I (DanConwayDev) don't think this is worth making this change as it
is breaking.

BREAKING CHANGE: NIP-34: replace NIP-10 style threading with NIP-22
for status events