NIP-34: add PR and PR update events
With the advent of Grasp, we can now do PRs without relying on a
single contributor-selected Git server, which they would have to
provision and could be rugged at any time.

Note how the commit ID is included in the event and any updates,
be it force pushes or additional commits, are added by supplementary
PR update events. This ensures history is maintained and Nostr events
remain the authority, rather than Git servers.

Grasp servers with proactive sync enabled will copy the commits to
ensure that they are available if the contributor-selected Git server
is unreliable or removed.

This also provides an alternative for large patches, which are too
large to be included within a signle nostr event.

It is also more supported by Git tools which don't always have patch
support. For example, isomorphic-git, which enables Git in the
browser.