NIP-34: send patches and issues to repos not yet on Nostr
Allow contributors to share patches and issues without requiring
maintainers to first create a Nostr repository announcement.

Motivation:

- Aligns with the permissionless philosophy of Nostr.
- Provides a way to contribute when the author doesn't have — or
  can't create — an account on the centralized git platform chosen by
  maintainers.
- Allows contributors who want a repository moved to Nostr to:
  1. Demo it to maintainers using their own data and community.
  2. Build community support by encouraging other community members
     to comment on, review, and interact with the issues/patches and
     to create their own issues/patches.

Once this community support is achieved and demonstrated and the repo
moves to Nostr, the first issues / patches shouldn't be disconnected
or unavailable. The `retired-clone` tag is intended to make it easier
for clients to find and display this content.